refactor: restructure project as pnpm monorepo
- Move existing backend code to packages/backend/
- Split dependencies between root (dev tools) and backend (runtime)
- Add pnpm workspace configuration with "backend" filter script
- Use root biome.json configuration for all packages
- Preserve all existing functionality without implementation changes
