+++ packages/backend/src/middleware/logger.ts
@@ -1,7 +1,7 @@
 import type { MiddlewareHandler } from "hono";
 import { logger } from "../utils/logger.js";
 
-const isProduction = process.env["NODE_ENV"] === "production";
+const isProduction = process.env.NODE_ENV === "production";
 
 export const tslogMiddleware = (): MiddlewareHandler => {
 	return async (c, next) => {
