+++ pnpm-workspace.yaml
@@ -1,3 +1,6 @@
+packages:
+  - 'packages/*'
+
 ignoredBuiltDependencies:
   - esbuild
